+++ Example/AcceptDemoTest/AppDelegate.h
@@ -1,0 +1,17 @@
+//
+//  AppDelegate.h
+//  AcceptDemoTest
+//
+//  Created by Danko, Radoslav on 24/02/2017.
+//  Copyright © 2017 radoslav.danko. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface AppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (strong, nonatomic) UIWindow *window;
+
+
+@end
+