+++ Example/AcceptDemoTest/ViewController.h
@@ -1,0 +1,15 @@
+//
+//  ViewController.h
+//  AcceptDemoTest
+//
+//  Created by Danko, Radoslav on 24/02/2017.
+//  Copyright © 2017 radoslav.danko. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface ViewController : UIViewController
+
+
+@end
+