+++ Example/AcceptDemoTest/ViewController.m
@@ -1,0 +1,29 @@
+//
+//  ViewController.m
+//  AcceptDemoTest
+//
+//  Created by Danko, Radoslav on 24/02/2017.
+//  Copyright © 2017 radoslav.danko. All rights reserved.
+//
+
+#import "ViewController.h"
+
+@interface ViewController ()
+
+@end
+
+@implementation ViewController
+
+- (void)viewDidLoad {
+    [super viewDidLoad];
+    // Do any additional setup after loading the view, typically from a nib.
+}
+
+
+- (void)didReceiveMemoryWarning {
+    [super didReceiveMemoryWarning];
+    // Dispose of any resources that can be recreated.
+}
+
+
+@end