+++ Example/AcceptDemoTest/main.m
@@ -1,0 +1,16 @@
+//
+//  main.m
+//  AcceptDemoTest
+//
+//  Created by Danko, Radoslav on 24/02/2017.
+//  Copyright © 2017 radoslav.danko. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+#import "AppDelegate.h"
+
+int main(int argc, char * argv[]) {
+    @autoreleasepool {
+        return UIApplicationMain(argc, argv, nil, NSStringFromClass([AppDelegate class]));
+    }
+}